moved test to other file
moved the tests of ui/__init__ to a new test file
